refactor(identity): ErrorCode enum 패키지 분리 #48
