+++ systems/identity/identity-domain/src/main/kotlin/hs/kr/entrydsm/identity/domain/exception/IdentityDomainException.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.identity.domain.exception
 
-import hs.kr.entrydsm.identity.domain.ErrorCode
+import hs.kr.entrydsm.identity.domain.enum.ErrorCode
 
 class IdentityDomainException(
     errorCode: ErrorCode,
